Backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Werkzeug==0.15.4
psycopg2-binary==2.8.5
SQLAlchemy==1.3.4
flask-wtf
WTForms-Components==0.9.7
flask-moment
babel

flask-debugtoolbar==0.10.0
Flask-Mail == 0.9.1
Flask-Login == 0.3.2